+++ blog/src/main/java/entity/User.java
@@ -4,7 +4,7 @@ package entity;
  * Created by scott on 2017/3/15.
  */
 public class User {
-    private Long id;
+    private Integer id;
 
     private String nickName;
 
@@ -18,13 +18,13 @@ public class User {
 
     private String idNo;
 
-    private Long roleId;
+    private Integer roleId;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -76,11 +76,11 @@ public class User {
         this.idNo = idNo;
     }
 
-    public Long getRoleId() {
+    public Integer getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Long roleId) {
+    public void setRoleId(Integer roleId) {
         this.roleId = roleId;
     }
 }
